// src/services/generated/models/roleType.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * SEMD API
 * OpenAPI spec version: 1.0.0
 */

export type RoleType = typeof RoleType[keyof typeof RoleType];


export const RoleType = {
  GUEST: 'GUEST',
  MEMBER: 'MEMBER',
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN',
} as const;